[Client] Manual synchronization now sets the config value as well
If the client is started without SSH host there is no synchronization happening. If the user
is now entering the host and manually synchronizes it would be insane to do the same again
when starting the client the next time. Therefore the manual synchronization sets the config
value upon the first synchronization.

Signed-off-by: Jim Martens <github@2martens.de>