[Client] Prevents overwrite of change in preferences dialog
Previously any change made to the settings would be overwritten
if a synchronize action was performed afterwards. That happened,
because a config variable would be set on the old config object
without the change and then this old object was written to disk -
overwriting the newer file in the process.

Signed-off-by: Jim Martens <github@2martens.de>